(1) The built-in function flatten-ands-in-lit is now a guard-verified :logic mode function.  (2) Installed update to :doc hints about :do-not-induct that I missed earlier.
A technical change related to (1) is that the ACL2 sources no longer
rely on books/system/kestrel.lisp to make dumb-negate-lit a
guard-verified :logic mode function.

Thanks to Alessandro Coglio for encouraging the enhancement, (1).

In addition to the usual "regression-everything" I successfully ran
"make proofs" and "make devel-check".

Regarding (2), the update to :doc hints about :do-not-induct: I
intended to commit this earlier, but I forgot.  Thanks to Mihir Mehta
for noticing and pointing that out.