Fixed makefile bug that was excluding models/y86/ books for exactly the wrong Lisps.  Fixed an excess of colon (:) characters in a json book.  Tweaked :doc for all-vars.  Changed raw-Lisp implementation of read-file-iterate-safe to avoid some stack overflows.  Fixed the v8-4 fix for print-gv, avoiding some very large expansion files.  Eliminated a compiler warning for Allegro CL during the build.
Thanks to Mihir Mehta for a Slack conversation leading to an
improvement to :doc system-utilities, clarifying that all-vars returns
a duplicate-free list.

The print-gv modification avoided a stack overflow when certifying
books/misc/redef-pkg.lisp using Allegro CL and some others, which was
caused by a large expansion file that contained an ACL2 world.

In books/centaur/fty/tests/, certification was failing when using
Allegro CL for the books deftranssum.lisp and deftagsum-scale.lisp,
because of stack overflows when attempting to read the
@useless-runes.lsp files using the existing ACL2 raw-Lisp
implementation of read-file-iterate-safe.  That implementation now
uses Common Lisp LOOP to avoid the stack overflow.  The new
implementation might be faster in some Lisps.

Regarding "eliminated a compiler warning for Allegro CL", quoting a
comment in (defxdoc note-8-4 ...):

; (Allegro CL only) Eliminated a compiler warning during the build for function
; from-to-by-ac, about unreachable code, by moving #+allegro code to
; with-warnings-suppressed from with-more-warnings-suppressed.