[APT] Improve the SOLVE transformation.
Locally generate a more general solution correctness theorem, along with an
instance of the target function that instantiates the function variable with the
solution function. This is done for the rewriting methods for now, but will be
soon done for the manual method as well. Then the theorem and the target
function instance will be generate non-locally.