[ATC] Extend accepted ACL2 code.
Now ATC accepts IF calls whose test is MBT or MBT$: it ignores test and 'else'
branch, and recursively processed/translates the 'then' branch.

These IF calls may arise from the APT::RESTRICT transformation, for
instance. This transformation may be used, in an APT derivation, to restrict the
range of certain ACL2 functions over the integers in a way that enable the
provably correct use of C bounded integer values (via transformations like
ISODATA and SIMPLIFY).