[Std/util] Add a new DEFTUTORIAL utility.
This generates macros useful to generate XDOC manual pages organized in a
sequence with links to 'next' and 'previous' pages. The user calls macros to
define the pages, in the desired order, and then a final macro all generates the
actual XDOC topics, automatically adding the appropriate next/previous links.

The name of this utility is motivated by the fact that this manual page
organization is typical of tutorials. The name could be generalized if there is
a good proposal for that.