[ATC] Improve proof generation.
Use the DEFOPENERS utility from [books]/kestrel/utilities/defopeners.lisp to
create opener rules for all the cases of the execution functions automatically.

The proofs for all the current tests work with these new opener rules. This is
currently an "alternative" set of openers to the ones created with DEFOPENER,
but we plan to swap their roles soon.