[ATC] Set up calculation of C return types.
So far we only support the C int type, so every generated C function has result
type int. However, when we support more C types, it is necessary to determine
the result type of each function. This will be calculated from the function
bodies, as part of their generation. This commit extends the ACL2 code that
translates ACL2 terms to C expressions and statements to also return the C type;
for now it is always int, but this will be generalized.