[ATC] Improve the calculation of types.
Build and use an alist from ACL2 variables to C types to determine the types of
variables. These are all int for now, but they will be generalized.

Associate a result type to each unary and binary operator, and use that as
result of a corresponding unary or binary expression. These types are all int
for now, but they will be generalized.