Changed "Source:" to ":DOC source:" on line 3 of each topic displayed in acl2-doc.
This change was discussed on the acl2-books list.  It was implemented
by a change in emacs/acl2-doc.el.  (I also made such a change in
books/xdoc/save-rendered.lisp, but as far as I can tell it didn't
affect anything; still, it seems harmless and appropriate.)  Thanks to
Mihir Mehta for a query leading to this change, and to Alessandro
Coglio and Eric Smith for helpful comments.