[Std/alists] Add theorem.
This is disabled by default, because the hypothesis is about alists while the
conclusion is about lists and involves a common predicate.

There is already a compound recognizer version of this rule, but in some cases
the rewrite rule seems needed. This new theorem can be enabled in those cases.

Another option could be to just add :rewrite to the rule classes of the existing
compound recognizer rule. Maybe there should be a more unified policy about
this, at least in the Std library, ideally as a more general ACL2 style guide.