[ecurve] Add predicate about prime field squares.
This is more general than elliptic curves; it probably belongs to the prime
fields library. But it is used for certain kinds of elliptic curves -- this is
why it is here now.

This non-constructively checks whether a prime field element is a square. The
witness function returns a square root, when the predicate holds.