[ecurve] Extend the formalization of Jubjub.
Expand the documentation of the a coefficient.

Prove that the d cofficient is not a square, using the just added theorem with a
weak form of Euler's criterion. Use the fast modular exponentiation operation
from the arithmetic-3 library to prove the key non-equality.

Prove that Jubjub is complete, since a is a square and d is not.