[ecurve] Add a formalization of Montgomery curves.
These are another form of elliptic curves. They are a special case of
Weierstrass curves (not short ones, but general ones).

The formalization includes a fixtype for these curves, a predicate to check if a
point is on a curve, neutral element, addition operation, negation operation,
and subtraction operation.

There is no closure proof of addition for now. This is needed to define scalar
multiplication in a guard-verified way. This will be eventually added.