[ecurve] Add a theorem about twisted Edwards curves.
This says that the only point on the curve with x = 0 or y = 1 are (0,1) and
(0,-1).

This is relevant to the birational equivalence with Montgomery curves: the
rational mappings do not work for points with x = 0 or y = 1, because they make
some denominators 0. This theorem says that we only need to worry about two such
points -- and map them specially, i.e. without the rational mapping.