[Isar] Start an Isar library.
Isar (= Intelligible Semi-Automated Reasoning) is Isabelle's structured proof
language, which is both machine-processable and human-readable. This library is
a (very small, for now) start towards providing a similar capability in ACL2,
which may be useful in certain circumstances, such as when programmatically
generating proofs according to handwritten templates, or when the proof
developer wants to explain why and how a proof works. It may be also useful to
guide the prover towards the right facts, as an alternative to :USE hints.

This does not introduce any new proof capabilities or axioms. It merely
generates ACL2 theorems according to the script. ACL2's built-in proof
capabilities are used inside the script, along with the ability to use
:HINTS. The granularity of the Isar proof steps can be as small or large as
desired. A single-step Isar proof is isomorphic to an ACL2 traditional proof.
In some cases, one can develop an Isar proof, and then incrementally automate
more of its steps via (sufficiently generic!) rules.

See the documentation for more discussion, and for a comparison with the ACL2
proof builder.