[ATC] Simplify the C dynamic semantics.
Have the mutually recursive function to execute (possibly non-pure) expressions
use the one for pure expressions after handling function calls. Eliminate the
mutually recursive function to execute lists of expressions, since we use the
one for pure expressions for function arguments.

After this change, one of the tests exhibited a proof failure that necessitated
a refinement of the generated proof strategy.