[ATC] Improve C static semantics.
Add an invariant requiring that the set of return types in a statement type is
always non-empty. This is because every statement must always terminate in some
way; note that the return types include NIL to describe execution that
terminates without actually returning a value (but continuing with the next
statement, if any).