Fixed merge-io-pairs and use it in books/kestrel/ethereum/top.lisp (where formerly that caused a certification failure).
Thanks to Alessandro Coglio for reporting this issue and to Eric Smith
for a key insight about the bug that led to its fix.

Technical note: The problem, as noted by Eric Smith, was that a
merge-io-pairs left the lenience as nil upon its conclusion, even if
the lenience was permitted coming into the merge-io-pairs call.  The
solution is to use a push/pop approach.