[ecurve] Extend the formalization of Montgomery curves.
Add nullary DEFUN-SKs that express the closure of addition and the associativity
of addition. Proving these properties is a bit of work (for the first one) and a
lot of work (for the second one), which we plan to do eventually, but in order
to make more progress in certain applications, for now we will use these as
additional assumptions where needed. Note that the rewrite rules associated with
these two DEFUN-SKs, which are of course universally quantified, corresponds
exactly to the desired theorems, but with the DEFUN-SK as additional hypothesis
(we use DEFINE-SK, which automatically generates the :DIRECT rewrite rule, when
that is a legal rewrite rule).

If these DEFUN-SKs turn out to be false, due to some (fixable) mistake in the
definition of addition or in their own formulation, any theorem with them as
hypotheses is logically vacuous. But this is still a much better situation than
being in an inconsistent logical theory, which is why we do capture these two
properties as axioms.