[ecurve] Improve specification of Montgomery curves.
Despite still not having the addition closure property proved, simplify the
definition scalar multiplication to just return a point, avoiding the flag
result that says whether the point is on the curve.

In order to verify the guards, the recursive auxiliary function for
multiplication by non-negative scalar still internally checks if the point
returned by the recursive call is on the curve (prior to adding it to the input
point), returning an irrelevant point if it is not. A separate theorem then
proves that the resulting point is always on the curve given the closure
assumption.

This approach was chosen to avoid having to put the nullary predicate that
expresses the closure assumption into the guard of scalar multiplication, and
thus of every function that depends on it, especially considering that the
predicate is not executable.

Adapt the definition of point order to the change.