[Semaphore] Add proofs of Semaphore components.
Each newly added .lisp file reads an R1CS circuit (in JSON format, from the
json/*.json files) and proves the equivalence of the circuit to its
specification. Both directions of the equivalence are covered; some of these
components require witness finding for the (specification ==> circuit) direction
of the proof.

The files contain explanatory comments about the proof methodology and some of
the proof details.