[ATC] Restructure implementation of event generation.
Instead of generating first the C translation unit and then the events
(constants and theorems), now we generate the events along with the ASTs that
form the translation unit.

The new structure is more appropriate for an upcoming extension of proof
generation.

It is also more appropriate for some possible future extensions in which we may
want to generate "prooflets" (i.e. small targeted proofs) for smaller ASTs than
C functions. That is, as ATC translates an ACL2 construct to a C construct, it
could also generate a proof for that piece of the translation. The purpose is,
of course, to make the proofs more controlled and efficient.