[ATC] Calculate execution steps.
Now the code generation functions also calculate a number of steps (measured in
terms of the limit argument passed to the big-step execution functions)
sufficient to execute the generated expressions and statements. This information
is also stored in the function table, and used when there are function calls.

These limit values will be used in an upcoming extension of proof generation.