Fix regression
Pull request #1234 introduced a name conflict with some Hardin 2020
workshop books. In each case, the new lemma was more general than the
old lemma of the same name, so I went ahead and removed the old
lemmas. I didn't add a "Mihir M. mod" comment as is often the
practice - it seemed out-of-place to take credit for fixing something
like this - but @david-s-hardin as the author may want to know that
this happened.