[ATC] Improve input validtion and event generation.
Now we check the freshness of the generated theorem names (which are determined
indirectly by the :CONST-NAME input) prior to starting to generate the code and
events, i.e. during the initial input validation phase. This results in better
error messages, and also potentially avoid the generation of local event names
that would conflict with the exported names.