Install-not-normalized now uses the untranslated body.  Assoc-equal-cadr and assoc-eq are built-in logic-mode functions.  Fixed an error message for :monitor.  Added new utility, logical-defun.  Deleted two obsolete short comments.
Changed install-not-normalized to use the untranslated body (and
updated its :doc and :doc note-8-4-books accordingly).  Thanks to Eric
Smith for the suggestion.  Also added a bit of error-checking to that
utility.  Note that some implementation functions are now :program
mode because of a program-mode function, logical-defun, now used in
the implementation.

Added guard to assoc-equal-cadr and put it into :logic mode.  Added
analogous function assoc-eq-cadr.  (The first remains undocumented and
the second also isn't documented.)

Made several small ACL2 changes, listed below.  The first bit below is
a comment added to (deflabel note-8-4 ...).

; Fixed the error message wording when attempting to monitor only simple
; (abbreviation) runes.  Thanks to Mihir Mehta for reporting the issue and
; providing a fix that we incorporated.

Added utility logical-defun (documented) in support of
install-not-normalized.

Deleted a couple of obsolete short source comments, in consultation
with J.