[ATC] Add model of unsigned integer operations.
These are all the unary and binary operations that apply to integers. Besides
the ones on the standard signed integer types that we had before, we now also
have the ones on the standard unsigned integer types. These are all generated by
macros, which captures the common structure of the different types.