[ecurve] Improve the fixtype of Montgomery curves.
Incorporate the primality requirement of the P component of the curve into the
definition of the curve fixtype (i.e. the :REQUIRE of DEFPROD). Eliminate the
separate primality predicate.

This is made possible by ADD-IO-PAIRS and its use for the primes for which we
create Montgomery curves. Otherwise, it would take too long just to construct
such curves (because ACL2 would execute RTL::PRIMEP); this is why the initial
definition had primality in a separate predicate, at a time before ADD-IO-PAIRS.

This improved fixtype of Montgomery curves results in various simplifications to
the functions and theorems that involve Montgomery curves, which no longer have
to deal with the separate primality predicate.

Adjust the birational equivalence with twisted Edwards curves to the change.

Adjust the definition of Zcash's Jubjub in Montgomery form to the change.