[ecurve] Prove commutativity of Montgomery addition.
This is one of the important group properties.

This is a bit of work to prove. There are several cases to consider, some of
which are straightforward, but the general case of finite points with distinct x
requires a bit more work. While it is not difficult to show that lambda (the
"slope") and x3 commute, it is not immediate for y3; however, a few algebraic
manipulations can turn y3 in a form whose commutativity can be established,
therefore establishing the commutativity in all cases.

This needed some additional prime field theorems, which have been put into a
separate file, along with others from before.