[ATC] Extend C static semantics.
Add support for all the pure binary operators, except for addition, subtraction,
and comparison of pointers.

The structure is also easily extensible as we add more types, e.g. things are
phrase in terms of arithmetic types for certain operators (not just integer
types, even though those are currently the only arithmetic types in our model).