[axe] Improve variable substitution.
Now we can use substitution to replace a var with a constant and drop the equality (previously rewriting would put in the constant and the equality would remain).

Now substitution may prove the entire clause, if a literal becomes a non-nil constant, and will drop literals that become the constant nil.