[ATC] Generate code and proofs for the recently added test.
This was previously causing the generated proof to be very slow or
nonterminating. With the extension to the symbolic execution theory made in the
previous commits, this theorem now works. However, it is a bit slow -- on the
order of tens of seconds, which is generally acceptable but seems unjusitified
for such a small example.

This suggests that we may need to generate more nuanced, modular, and controlled
proofs fairly soon, also because the "debugging cycle" for this kind of issues
is not particularly efficient.