[ATC] Withdraw support for MBT/MBT$ tests for conditional expressions.
MBT/MBT$ tests are still supported for (ACL2 terms that represent) conditional
statements, but not for (ACL2 terms that represent) conditional expressions any
more. This simplifies implementation and documentation, and avoids some
duplication. Note that, in a derivation, one can always lift IFs to outer levels
so that they represent C statements rather than C expressions, and still get the
same desired effect, namely to "replace" (IF (MBT/MBT$ ...) TERM ...) to TERM.
We do not expect this pattern to occur in many places, and therefore to
significantly increase the size of terms. In fact, the main use case is of such
a pattern at the outermost level of a function body, resulting from the RESTRICT
transformation or other transformation that produces a similar wrapping. (If
size explosion turns out to be a problem later, we can consider re-supporting
MBT/MBT$ for conditional expressions again.)