[ATC] Add explicit support for conditional expresions.
We add an identity function called CONDEXPR that, when put around an IF, tells
ATC to generate a C conditional expression instead of a C conditional statement.
Before, the choice was based on where the IF occurs, which in some cases was the
only choice, but in other cases both choices could have been equally possible,
based on preference and style. So, instead of incorporating heuristics in ATC,
and in accordance with ATC's philosophy of delegating the exact form of the C
code to external tools like APT, we now require the use of (CONDEXPR (IF ...))
to represent a conditional expression ... ? ... : ...; any unwrapped IF
represents a conditional statement if (...) { ... } else { ... }, instead, which
is by far the more common case of the two.

Slightly extend the symbolic execution theory to expand CONDEXPR -- obviously.

Adapt a few tests as necessary.