[ATC] Improve the generation of the integer operations.
Add and use a specialized function definition macro for the ACL2 functions that
model C integer operations and their guards.

This not only makes the definitions more concise, but also reduces the times to
process and include the file by roughly half. This is because they avoid
generating some table events, generated by DEFINE (which was previously used,
providing almost comparable conciseness).

Thanks to Eric Smith for a useful discussion on the book inclusion time.