[ATC] Improve/fix treatment of LETs and MV-LETS.
The same checks that were applied to assignments in the preceding bug fixes must
be also applied to transformations (which are like generalized assignments).

The relevant ATC code has been refactored to make all the checks more clear and
explicit.

The relevant ATC user documentation has been refactored, in particular
introducing a notion of 'assignable' variable.