Include-book-dir, flatten-ands-in-lit!, union-equal-to-end, and strip-non-hidden-package-names are now guard-verified :logic mode functions.  New utility: Get-guard-checking.  Fixed invariant involving raw-include-book-dir-p.  Fixed brr-prompt for script-mode in ACL2(r).  Tweaked :doc for evisc-table and CCL installation.  Tweaked ev-fncall-rec-logical.
Quoting :doc note-8-4:

  A new utility, [get-guard-checking], returns the [guard]-checking
  value most recently installed, either t (when the ACL2 executable
  was built) or presumably by [set-guard-checking].  Thanks to Eric
  McCarthy for suggesting this utility.

The issue with an invariant involving raw-include-book-dir-p came up
when considering the guard for include-book-dir: state global
'raw-include-book-dir-alist has value :ignore if and only
'raw-include-book-dir!-alist does.  The failure of that invariant
was probably ephemeral and benign, but now it's fixed.

Script-mode changes the ACL2(r) prompt to be the same as the ACL2
prompt, which supports the use of the tool run-script in the
regression.  This change was not being done inside brr.  Thanks to
Jagadish Bapanapally for reporting an ACL2(r) regression failure of
books/demos/brr-free-variables-book.lisp

Thanks to Eric Smith for suggesting the :doc tweaks and the conversion
of include-book-dir and strip-non-hidden-package-names to
(guard-verified) :logic mode.

Passed "make proofs" and "make devel-check" in addition to
the usual "regression-everything".