Updated "bleeding-edge" manual URL in installation instructions.
The old URL, which still works and should continue to do so, was:

http://www.cs.utexas.edu/users/moore/acl2/manuals/current/manual/

The new URL is:

http://www.cs.utexas.edu/users/moore/acl2/manuals/latest/

This new URL was added on the web on 6/1/2021, actually as a soft link
to the old URL, which still works and should continues to do so.
Thanks to Eric Smith for discussions leading to that change.