[ATC] Improve support for local variables.
Now ATC supports (and in fact requires) the use wrappers DECLAR and ASSIGN to
distinguish between LETs that represent local variable declarations and LETs
that represent local variable (or function parameter) assignments. The wrappers
are identities, and must be put around the term to which the LET variable is
bound. A LET without either wrapper represents the transformation of a single
variable via statements, i.e. neither (just) a declaration or an assignment.
MV-LET still takes no wrappers, because it is only used to represent
transformations of multipe variables via statements.

See :DOC ATC-LET-DESIGNATIONS for an explanation of the rationale behind this
approach.

Adapted the symbolic execution theory to enable both wrappers.

Updated the user documentation.

Updated the tutorial pages that describe local variable declarations and
assignments.

Adapted the existing tests by adding the wrappers.