[ecurve] Improve formalization of twisted Edwards curve.
Add the primality requirement to the invariant of the curve fixtype, and remove
the separate predicate expressing that primality requirement.

The separate predicate was justified before the introduction of the ADD-IO-PAIRS
utility, but now that is no longer necessary.

This simplifies several guards and theorems.

Adapt files that use twisted Edwards curves to this change.