[ecurve] Add twisted Edwards associativity assumption.
This is analogous to what was done for Montgomery curves.

We introduce a nullary DEFUN-SK that asserts the associativity of twisted
Edwards addition. This DEFUN-SK can be be used as hypothesis in theorems whose
proofs rely on twisted Edwards associativity.

We do this because elliptic curve associativity proofs are reportedly laborious,
so we do not want this to be in the critical path.