[Std/system] Generalize the CHECK-MV-LET-CALL utility.
Now this utility also checks for the presence of HIDE wrappers in the term,
which are introduced when translating an MV-LET that includes IGNORE
declarations. The utility now returns a mask of booleans saying which wrappers
are present. See the (updated) documentation of the utility for details.

Thanks to Matt Kaufmann for useful discussion.