[json-parser] Attempt to add full Unicode support.
Characters encoded in UTF-8 should pass through the parser unchanged.  But now we also handle Unicode escapes, both single escapes for characters in the Basic Multilingual Plane, and double escapes representing UTF-16 surrogate pairs.

Thanks to Eric McCarthy for requesting improved Unicode support and for providing a helpful pull request that inspired parts of this implementation.

It would be good to add more tests of the new functionality.