[ATC] Simplify a test.
Now ATC supports MV-LETs with IGNORE declarations, thanks to the recent
extension of the CHECK-MV-LET-CALL utility.

Thus, it is no longer necessary to do (SET-IGNORE-OK T) before a target function
that includes MV-LETs that need to ignore some bound variables. Instead, we can
use IGNORE declarations directly in the MV-LET.