[ATC] Extend the symbolic execution theory.
Add two rules to discharge the recurring hypotheses that the limit is not 0
(i.e. that the symbolic execution has not run out of the artificial counter).
These apply to the case in which the limit is not a constant, which is how we
are generalizing the symbolic execution proofs.