[ATC] Generalize symbolic execution proofs.
Now we prove the functional correctness theorems with variable limits directly
by symbolic execution, without going through lemmas that use constant limits.
These lemmas are no longer generated, in fact. It still remains to remove the
code that constructs those lemmas, and to merge its dev doc with the one for the
modified code that generates the correctness theorems; this will be done in an
upcoming commit.