Added new :doc topic, developers-guide-pitfalls.
This is mostly just a stub for readers of the Developers Guide
(presumably potential future ACL2 developers), as a place to
list some things to watch out for.  So far it just says to be
careful not to ignore tag trees that may contain assumptions.

Thanks to Pete Manoilios and Eric Smith for comments leading to the
addition of this topic.