Fixed bug reported in GitHub Issue #1275 (evident in SBCL but not in CCL).
The bug was in git commit 00813021f04d953d45726c097c937c67a5113130,
specifically pertaining to the item, "Fixed raw-mode to do a better
job maintaining global stobj values."  Thanks to Pete Manolios and
Andrew Walter for reporting the bug in GitHub Issue #1275 and on
Slack, respectively.

The bug didn't show up in CCL, which is what I had used for testing.
To see why, evaluate the following in CCL and in SBCL:

(let ((x 3) (x 4)) x)

Actually CCL prints a warning, which however is suppressed in the ACL2
loop (perhaps because "ignore" warnings are suppressed in the loop, in
support of set-ignore-ok).