[ATC] Improve generated theorems.
Use (IDENT <quoted-string-constant>) as the canonical representation of C
identifiers, instead of exposing their internal representation as quoted
identifier constants. This was actually already the case for the exported
theorems, but it now also the case for the local theorems. The symbolic
execution has been modified to use that canonical representation as well, so
that the theorems in the new form can be compositionally used as rewrite rules.

This also necessitated additional symbolic execution rules to handle function
environments. We plan a more general handling of function environments
eventually anyways.