[ATC] Generate additional events for loops.
We generate a type prescription rule for each recursive target function,
asserting that the measure yields a natural number. This is the same as the
corresponding applicability condition (recently introduced), but we need it to
be a type prescription rule here. We extend the function information with the
name of this theorem for each recursive function, as it needs to be used in
upcoming generated theorems.