[ATC] Generate more events for loops.
This is like the termination theorem of the ACL2 function that represent the
loop, but with O< replaced with <, which relies on the fact that the measure
yields a natural number, as checked via applicability conditions.

This variant of the termination theorem serves to establish the hypotheses of
the induction hypothesis in the main loop correctness termination theorem, which
will be generated soon.