[ATC] Generate the main loop correctness theorem.
This works on some examples, but not yet on others. Some of the issues have been
identified and will be addressed soon.

Also added an undocumented :DEVELOPER (T/NIL) option to have ATC generate these
theorems. This is means as a developer option to ease development and debugging;
it is not for users.