Speed up building of the manual.
Quoting :doc note-8-4-books:

  Building of the combined ACL2+Books manual (by certifying
  doc/top.lisp) has been made more efficient (more than 20% reduction
  measured when such changes were made).

The key changes were as follows to reduce the time from 302 seconds to
237 seconds on my laptop (perhaps imprecise measurements, both because
it's a Mac and because browser was running, with music).  That's just
over 5 minutes to just under 4 minutes, hence the claim above of "more
than 20% reduction".

- Use princ$ instead of fms! to print the acl2-doc-search file --
  saved about 50 seconds

- Use Eric Smith's version of Sol Swords's archiving trick for the
  ACL2s documentation, with a no_port comment in new file
  books/acl2s/top-doc.lisp to prevent incorporating books into the
  portcullis -- probably saved about 13 seconds.  (Thanks to Eric for
  noting that ACL2s files were probably the main bottleneck in loading
  the documentation.)

It looks like we could save about 19 more seconds by skipping the
writing of files that support the acl2-doc Emacs-based documentation
browser: rendered-doc-combined.lsp and acl2-doc-search.  These are
automatically downloaded from the internet as needed when using
acl2-doc, so maybe skipping these is worthwhile by default; but 1/3
minute out of 4 minutes isn't a lot so I haven't bothered to skip
them.  If there's interest in doing so, I could add an "opt-in"
environment variable.

By the way, I compared the new and old versions of
books/system/doc/rendered-doc-combined.lsp and it looks like the
changes have no functional impact -- just efficiency.