[ATC] Improve proof generation for loops.
We now generate a function for the measure of each recursive target function, so
that the generate proofs can treat the measure as a black box and are not
thwarted by possibly modifications of the measure terms, e.g. due to case
analysis in proofs.

However these measure functions cannot quite be local: they have to be exported
because they are mentioned in the generated theorems, precisely in the lower
bounds of the limit variable of the big-step operational semantics. Perhaps a
better course of actions is to "move" the creation of these measure functions to
a new APT transformation, and then have ATC require that recursive target
functions have measures of the form (<fn> <subset-of-formals>); this approach
seems in line with the synergy between APT and ATC that is the basis for the
design of ATC.