[ecurve/primes] Move XDOC topic content.
There was a duplication between an (empty) XDOC topic created by DEFPRIME and an
explicit DEFXDOC with the same name. The DEFXDOC was created before DEFPRIME has
XDOC support. Now the content of the DEFXDOC has been moved into the DEFPRIME
call, eliminating the DEFXDOC.

Thanks to Matt Kaufmann for reporting the issue.