[ATC] Improve the C static semantics.
Use the (newly introduced) type void, instead of NIL, in the sets of types
returned by the static semantics functions that check statements. The void type
indicates that the statement in question may return nothing. This lets us use
sets of types instead of sets of optional types in the statement type records.