[FTY] Add DEFRESULT and supporting utilities.
This macro generates types consisting of good results (specified as some
existing fixtype) and error results of type RESULTERR, which is provided along
with the macro. The macro is accompanied by a B* binder to concisely check and
propagate errors, analogous to the ER binder (which expands to ER-LET*). The
propagation also constructs a stack of function names corresponding to the call
stack; it relies on the current function name being assigned to __FUNCTION__,
which is automatic when using DEFINE. The stack and error value is initialized
via an ERR macro. This way, when an error is returned, besides the core
information there, (the function naems in) the call stack is also visible.