[ATC] Include FTY library for pseudo-terms.
This commit is just to make sure that including the library doesn't cause
anything to fail (e.g. because of newly introduced rules).

The plan is to switch ATC to systematically use FTY operations on pseudo-terms,
because currently processing certain ATC functions with the built-in operations
on terms is slow.