[ATC] Extend proof generation.
The theorems saying that the target functions return C values need to expand the
definitions of the shift operations to the versions that take ACL2 integers as
shift distances. This is analogous to what's done in the proofs of the
correctness theorems.

Thanks to Jim McDonald for finding a failing test that uncovered the need for
this extension.