[ATJ] Fix bug in pre-translation.
The type annotation step of the pre-translation phase was failing to add the
proper conversion around ACL2 terms representing Java array constructions with
initializers. It was adding an identity conversion, failing to take the required
(i.e. destination) types (when present) into account. This was presumably due to
a typo, or copy-and-paste error, since the pattern should be the same as for
other terms.

Thanks to Eric Smith for reporting a test that caused the generated Java code
not to compile due to a type mismatch between the expression returned by the
Java method and the return type of the Java method. That happened because the
required return type was not making it into the conversion wrapping the return
expression. Now that example correctly fails, because there is no support to
convert Java arrays to (Java representations of) ACL2 values. By assigning an
array return type via ATJ-MAIN-FUNCTION-TYPE, the code is generated and
compiles. (The call of ATJ-MAIN-FUNCTION-TYPE must specify an unnamed array as
output, because the array is created in the method, as opposed to being received
as an input and being modified by the method.)