Avoid Lisp-dependent failures, in part by making :OOPS output of OBSERVATION type.
For a couple of tests that compare resulting with expected output
logs, test failures showed up using LispWorks vs. CCL/SBCL because of
differences in what's printed when compiling on-the-fly or not.  Part
of the change involved changing :OOPS output to be controllable as an
OBSERVATION output type, as per this from :doc note-8-5.

  Output from :[oops] may now be inhibited as OBSERVATION output, by
  using [set-inhibit-output-lst] or [with-output].