[ATC] Extend the symbolic execution theory.
Add rules to expand SHL/SHR-type-OKP with type in {SCHAR, UCHAR, SSHORT,
USHORT}, which should have been there in the first place but were forgotten
somehow.

Also discover the need to add a rule to simplify IFIX away (which arises in
definitions of shift operations), in combination with some return type rules
that have also been added to the theory.

Thanks to Jim McDonald for reporting a failure that led to the discovery of
these issues.