[ATC] Speed up loop proofs.
Now the loop correctness theorem makes use of the loop body correctness theorem,
as a rewrite rule. This should be faster, in particular avoiding repeated
symbolic execution for at least the loop body statement.

This also paves the way for a simpler induction proof for loops, once we also
have a separate proof for the loop test expression.