[Zcash] Add Theorem A.3.5 from the Zcash specification.
This is a critical theorem for the correctness of the Pedersen gadget.

This is not straightforward to mechanize in ACL2, particularly given the way in
which it is formulated and proved in the Zcash specification. This ACL2 proof
uses a different approach than the one in the Zcash specification. See the
documentation comments in the file for details.