[ATC] Speed up generated proofs.
In general, the proofs of the theorems for a target function FN use theorems
about the functions called by FN. Before this commit, the generated hints used
the theorems of all the preceeding target functions, whether they were actually
called by FN or not, which is fine for small examples. This commit only includes
hints for the functions that are actually called, which speeds up larger
examples.

Thanks to Jim McDonald for running sufficiently large tests that they made this
change more urgent.