[FTY] Reduce consing in recognizers for defprod, etc.
This avoids executing strip-cars every time a structure created by defprod is checked.  Profiling showed that the call to strip-cars was allocating the majority of the memory in a large proof that checked well-formedness of many (tens of millions) of small structures created with defprod.  This change also reduces the number of passes through the alist.  This change uses MBE and so effects execution only, not the logical definitions generated.