[ATC] Optimize generated proofs.
We now define a use an ACL2 theory for the symbolic execution rules. Since there
are thousands of them, this speeds up the execution of ACL2's hints translation.

Thanks to Jim McDonald for creating and running tests with thousands of C
functions, and to Eric Smith for profiling the execution and discovering that a
lot of time was spent in translating hints and for suggesting the definition and
use of a theory to speed that up.