[Yul] Ensure that all FTY::DEFPRODs have tags.
This may cause double tags when product values are used in single-field tagged
summands, but (1) it is better for all product values to carry a tag/type, and
(2) it should be possible to extend FTY::DEFTAGSUM to avoid the double tagging,
or have a more flexible FTY::DEF...SUM that is a mix of FTY::DEFTAGSUM and
FTY::DEFFLATSUM, i.e. it selectively adds tags to only certain summands.