[prime-fields] Improve inv.
Now inv of zero is always 0, regardless of p.  This is based on a suggestion from Alessandro Coglio.  Previously, (inv 0 2) was 1, based on the simple definition that relies of Fermat's Little Theorem.  But since inversion is not defined on 0, we can choose whatever behavior we want.  With the change, a few rules get slightly worse (added hyps), but some get significantly better.