[ATC] Generalize dynamic semantics.
Now the dereferencing operation on pointers operates on computation states
rather than heaps. This open the door to generalizing the model to include
pointers to non-heap memory regions.

Another benefits is that the for of the symbolic computation states, and the
formulation of the correctness theorems, is now a bit simpler and more
streamlined.

The symbolic execution rules have been adapted accordingly.