Restored invariant that every function's well-founded-relation is indeed known by ACL2 to be well-founded.  Fixed "make proofs" and "make devel-check".
Removed all but one "Do$ Wart" comment, restoring the invariant that
for every function's well-founded-relation is indeed known by ACL2 to
be well-founded.  (We build in that fact without proof but check it
with our "devel-check" process that is also used for justifying
admission of some built-in functions as guard-verified :logic mode
functions.)  Thanks to Eric Smith for a query leading to this
improvement.

Fixed two problems with existing ACL2 (before the changes above):

- Fixed inability to do "make proofs" caused by introduction of
  function retract-stobj-tables in support of stobj-tables.

- Fixed failure of "make devel-check" (see comment in source
  definition of *system-verify-guards-alist*) by modifying a
  proof-builder instruction in books/system/apply/apply-prim.lisp.
  Perhaps there is a better solution for that issue.

Passed "make devel-check" and "make proofs" in addition to the usual
"make regression-everything".  Note that none of the above issues were
present, as far as I know, in ACL2 8.4 (hence, there are no new
release notes).