[ATC] Generalize the generated correctness theorems.
These are for C functions (vs. C loops). The final computation state mentioned
in the theorem is no longer the same as the initial one, but rather it is a nest
of array writes that reflect side effects. Also, the result of EXEC-FUN is no
longer necessarily the ACL2 function's result, but it may be NIL if the C
function is void.