[ATC] Add distinct-pointer hypotheses to generated theorems.
Since C code that manipulates arrays (via pointers) is represented by ACL2 code
that manipulate arrays as a whole, correctness critically relies on the fact
that the pointers are distinct. The C model upon which the proofs rely already
ensures that arrays with different pointers do not overlap, which is another
assumption that calling C code must satisfy.